Removed faction limitation on production queues on Bio-Center, Robot Factory and Teleport
This is necessary because if a player with different faction (than defined in the ProductionQueue trait) had these buildings from game start (i.e. placed on map), they wouldn't be able to build anything from it. This is because the queue is disabled, if the player's faction isn't in the ProductionQueue.Factions list.