Make upgraded Big Eye require Heavy Rockets technology
The upgraded Big Eye requires the Light Plasma Cannon, but

Since upgrades are not implemented yet, the upgraded defense towers can be built straight from Constr/Prod. Centers. The upgraded Big Eye requires the Light Plasma Cannon, but it should not be possible to acquire it until the regular Big Eye can be produced. As such, requiring Heavy Rockets technology is necessary.

on Heavy Rockets technology